fixed adding and editing menu items

+++ src/UthandoNavigation/Service/MenuItem.php
@@ -78,9 +78,9 @@ class MenuItem extends AbstractService
 		$menuItem = $this->getById($model->getMenuItemId());
 	
 		if ($menuItem) {
-			// if page postion has changed then we need to delete it
+			// if page position has changed then we need to delete it
 			// and reinsert it in the new position else just update it.
-			if ($post['position']) {
+			if ($post['position'] && $post['menuInsertType'] != 'noInsert') {
 				// TODO find children and move them as well.
 				$del = $this->delete($model->getMenuItemId());
